remove the cvs jobmaster to the cvs job folder
